svc-ai-orchestrator/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.3
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.11.5
pydantic-settings==2.8.1
python-dotenv==1.1.0
python-jose[cryptography]==3.3.0
httpx==0.28.1

# LangGraph + LangChain (minimal)
langgraph==0.4.1
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-anthropic==0.3.12
langchain-google-genai>=2.0

# OpenAI + Anthropic direct SDKs
openai==1.82.1
anthropic==0.52.0

# Web search
tavily-python==0.5.0
